Move getHealthz code on separate script
