+++ apps/nextjs-website/src/lib/chatbot/chatbotApi.ts
@@ -98,34 +98,6 @@ export const getQueries = (query: string) =>
     )
   );
 
-export const getHealthz = () =>
-  pipe(
-    R.ask<ChatbotEnv>(),
-    R.map(({ config: { CHATBOT_HOST: chatbotHost }, getAuthToken, fetch }) =>
-      pipe(
-        // handle any promise result
-        TE.tryCatch(() => getAuthToken(), E.toError),
-        TE.chainTaskK(
-          (authToken) => () =>
-            fetch(`${chatbotHost}/healthz`, {
-              method: 'GET',
-              headers: {
-                'Content-Type': 'application/json',
-                Authorization: `Bearer ${authToken}`,
-              },
-            })
-        ),
-        TE.chain((response) => {
-          if (response.status === 200) {
-            return TE.tryCatch(() => response.json(), E.toError);
-          } else {
-            return TE.left(makeError(response));
-          }
-        })
-      )()
-    )
-  );
-
 export const getCurrentSession = () =>
   pipe(
     R.ask<ChatbotEnv>(),
